Feat #116: Add countries
In order to store information about cities that workshop is in
We need to store country information

Country code is used as id, names are in English but for now that is
going to be good enough. Since we need to prioritize some countries (in
order to show them more relevant when selecting) there is priority
defined too.